Move WAL delta logic into separate module
