+++ lib/collection/src/lib.rs
@@ -15,6 +15,7 @@ pub mod shards;
 pub mod telemetry;
 mod update_handler;
 pub mod wal;
+pub mod wal_delta;
 
 #[cfg(test)]
 mod tests;
