+++ lib/collection/src/shards/shard.rs
@@ -163,10 +163,11 @@ impl Shard {
 
     pub async fn resolve_wal_delta(&self, recovery_point: RecoveryPoint) -> CollectionResult<u64> {
         let resolve_result = match self {
-            Self::Local(local_shard) => local_shard.resolve_wal_delta(recovery_point).await,
+            Self::Local(local_shard) => local_shard.wal.resolve_wal_delta(recovery_point).await,
             Self::ForwardProxy(proxy_shard) => {
                 proxy_shard
                     .wrapped_shard
+                    .wal
                     .resolve_wal_delta(recovery_point)
                     .await
             }
