+++ lib/collection/src/wal.rs
@@ -2,21 +2,13 @@ use std::fmt::Debug;
 use std::marker::PhantomData;
 use std::path::Path;
 use std::result;
-use std::sync::Arc;
 use std::thread::JoinHandle;
 
 use io::file_operations::{atomic_save_json, read_json};
-use parking_lot::Mutex as ParkingMutex;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
-use tokio::sync::Mutex;
 use wal::{Wal, WalOptions};
-
-use crate::operations::OperationWithClockTag;
-use crate::shards::local_shard::clock_map::{ClockMap, RecoveryPoint};
-
-pub type LockedWal = Arc<ParkingMutex<SerdeWal<OperationWithClockTag>>>;
 
 #[allow(clippy::enum_variant_names)]
 #[derive(Error, Debug)]
@@ -271,53 +263,6 @@ impl<'s, R: DeserializeOwned + Serialize + Debug> SerdeWal<R> {
 
     pub fn segment_capacity(&self) -> usize {
         self.options.segment_capacity
-    }
-}
-
-/// A WAL that is recoverable, with operations having clock tags and a corresponding clock map.
-pub struct RecoverableWal {
-    pub(super) wal: LockedWal,
-    /// Map of all last seen clocks for each peer and clock ID.
-    pub(super) last_clocks: Arc<Mutex<ClockMap>>,
-}
-
-impl RecoverableWal {
-    pub fn from(wal: LockedWal, last_clocks: Arc<Mutex<ClockMap>>) -> Self {
-        Self { wal, last_clocks }
-    }
-
-    /// Write a record to the WAL but does guarantee durability.
-    pub async fn lock_and_write(&self, operation: &mut OperationWithClockTag) -> Result<u64> {
-        // Update last seen clock map and correct clock tag if necessary
-        if let Some(clock_tag) = &mut operation.clock_tag {
-            // TODO:
-            //
-            // Temporarily accept *all* operations, even if their `clock_tag` is older than
-            // current clock tracked by the clock map
-
-            // TODO: do not manually advance here!
-            let _operation_accepted = self
-                .last_clocks
-                .lock()
-                .await
-                .advance_clock_and_correct_tag(clock_tag);
-
-            // if !operation_accepted {
-            //     return Ok(UpdateResult {
-            //         operation_id: None,
-            //         status: UpdateStatus::Acknowledged,
-            //         clock_tag: Some(*clock_tag),
-            //     });
-            // }
-        }
-
-        // Write operation to WAL
-        self.wal.lock().write(operation)
-    }
-
-    /// Get a recovery point for this WAL.
-    pub async fn recovery_point(&self) -> RecoveryPoint {
-        self.last_clocks.lock().await.to_recovery_point()
     }
 }
